INT 241 create file for buildTs2741NextJsImageNoAltReplacementEnvelope
