+++ src/features/repairCode/buildReplacementEnvelope.ts
@@ -1,4 +1,3 @@
-import { createPrinter, createSourceFile, EmitHint, factory, JsxSelfClosingElement, NewLineKind, ScriptKind, ScriptTarget } from 'typescript';
 import { CaseKind } from '../../cases/types';
 import type { Classification } from '../../classifier/types';
 import type { ReplacementEnvelope } from '../../components/inferenceService';
@@ -14,79 +13,8 @@ import {
 	buildReplacement,
 	buildTs2769ObjectAssignReplacement,
 } from './buildReplacement';
+import { buildTs2741NextJsImageNoAltReplacementEnvelope } from './buildTs2741NextJsImageNoAltReplacementEnvelope';
 import { extractKindsFromTs2345ErrorMessage } from './extractKindsFromTs2345ErrorMessage';
-
-export const buildTs2741NextJsImageComponentMissingAttributeInferenceJob = (
-	file: File,
-	node: JsxSelfClosingElement,
-): ReplacementEnvelope => {
-	const sourceFile = createSourceFile(
-		'index.ts',
-		'',
-		ScriptTarget.Latest,
-		false,
-		ScriptKind.TS,
-	);
-
-	const printer = createPrinter({ newLine: NewLineKind.LineFeed });
-
-	const jsxAttribute = factory.createJsxAttribute(
-		factory.createIdentifier('alt'),
-		factory.createStringLiteral('', true),
-	);
-
-	const { properties } = node.attributes;
-
-	const lastAttribute = properties[properties.length - 1];
-
-	if (!lastAttribute) {
-		const jsxSelfClosingElement = factory.updateJsxSelfClosingElement(
-			node,
-			node.tagName,
-			node.typeArguments,
-			factory.createJsxAttributes(
-				[
-					jsxAttribute,
-				]
-			)
-		)
-	
-		const replacement = printer.printNode(EmitHint.Expression, jsxSelfClosingElement, sourceFile);
-
-		const start = node.getStart();
-		const end = node.getEnd();
-
-		const range = buildIntuitaRangeFromSimpleRange(
-			file.separator,
-			file.lengths,
-			{ start, end },
-		);
-
-		return {
-			range,
-			replacement,
-		};
-	}
-
-	const start = lastAttribute.getEnd();
-	const end = start;
-
-	const range = buildIntuitaRangeFromSimpleRange(
-		file.separator,
-		file.lengths,
-		{ start, end },
-	);
-
-	const width = lastAttribute.getLeadingTriviaWidth();
-	const triviaText = lastAttribute.getFullText().slice(0, width);
-
-	const replacement = printer.printNode(EmitHint.Unspecified, jsxAttribute, sourceFile);
-
-	return {
-		range,
-		replacement: `${triviaText}${replacement}`,
-	};
-}
 
 export const buildReplacementEnvelope = (
 	file: File,
@@ -133,7 +61,7 @@ export const buildReplacementEnvelope = (
 	}
 
 	if (classification.kind === CaseKind.TS2741_NEXTJS_IMAGE_COMPONENT_MISSING_ATTRIBUTE) {
-		return buildTs2741NextJsImageComponentMissingAttributeInferenceJob(
+		return buildTs2741NextJsImageNoAltReplacementEnvelope(
 			file,
 			classification.node,
 		);
